物体が認識できなかった時はLabelを出力しないようにし、identifierをnilにするようにした。 audio系の処理がViewに入っていたのでControllerに移動した
