+++ EnglishApp/View/ARView.swift
@@ -1,7 +1,6 @@
 import Foundation
 import UIKit
 import ARKit
-import AVFoundation
 
 protocol ARViewDelegate {
     func tapGesture(identifier: String?)
@@ -16,16 +15,11 @@ class ARView: UIView, ARSKViewDelegate, ARSessionDelegate, ObjectDetectionModelD
     let objectDetectionModel: ObjectDetectionModel
     let configuration: ARConfiguration
     var objectLabel: UILabel!
-    var speechSynthesizer : AVSpeechSynthesizer!
-    var audioPlayer: AVAudioPlayer!
-    let audioCorrect = NSDataAsset(name: "correct1")
-    let audioIncorrect = NSDataAsset(name: "incorrect1")
 
     override init(frame: CGRect) {
         self.sceneView = ARSKView()
         self.objectDetectionModel = ObjectDetectionModel()
         self.configuration = ARWorldTrackingConfiguration()
-        self.speechSynthesizer = AVSpeechSynthesizer()
         
         super.init(frame: frame)
         
@@ -88,19 +82,6 @@ class ARView: UIView, ARSKViewDelegate, ARSessionDelegate, ObjectDetectionModelD
         maru2.textAlignment = .center
         maru2.tag = 0
         self.addSubview(maru2)
-        
-        // AVAudioPlayerのインスタンスを作成,ファイルの読み込み
-        do {
-            audioPlayer = try AVAudioPlayer(data: audioCorrect!.data, fileTypeHint: "mp3")
-            audioPlayer.volume = 0.1
-        } catch {
-            print("AVAudioPlayerインスタンス作成でエラー")
-        }
-        // 再生準備
-        audioPlayer.prepareToPlay()
-        
-        // 再生する
-        audioPlayer.play()
     }
     
     func setWrongLabel() {
@@ -123,19 +104,6 @@ class ARView: UIView, ARSKViewDelegate, ARSessionDelegate, ObjectDetectionModelD
         batsu2.textAlignment = .center
         batsu2.tag = 0
         self.addSubview(batsu2)
-
-        // AVAudioPlayerのインスタンスを作成,ファイルの読み込み
-        do {
-            audioPlayer = try AVAudioPlayer(data: audioIncorrect!.data, fileTypeHint: "mp3")
-            audioPlayer.volume = 0.1
-        } catch {
-            print("AVAudioPlayerインスタンス作成でエラー")
-        }
-        // 再生準備
-        audioPlayer.prepareToPlay()
-        
-        // 再生する
-        audioPlayer.play()
     }
     
     func startSession() {
@@ -165,36 +133,34 @@ class ARView: UIView, ARSKViewDelegate, ARSessionDelegate, ObjectDetectionModelD
             view.removeFromSuperview()
         }
         
-        let objectRoundedLineView = UIView(frame: objectBounds!)
-        objectRoundedLineView.tag = 10
-        objectRoundedLineView.backgroundColor = UIColor(rgba: 0xf0ea9e)
-        self.addSubview(objectRoundedLineView)
-        self.subviews.filter({$0 is UILabel}).forEach { label in
-            self.bringSubviewToFront(label)
+        if let objectBounds = objectBounds {
+            
+            let objectRoundedLineView = UIView(frame: objectBounds)
+            objectRoundedLineView.tag = 10
+            objectRoundedLineView.backgroundColor = UIColor(rgba: 0xf0ea9e)
+            self.addSubview(objectRoundedLineView)
+            
+        } else {
+            print("Object detection is failed.")
         }
         
         self.identifier = identifier
-        
         // 物体の名前を中心に表示
         self.objectLabel.text = self.identifier
         self.objectLabel.font = UIFont(name: "Menlo", size: 100)
         self.objectLabel.frame = CGRect(x: 0, y: 0,width: self.frame.width, height: self.frame.height * 0.2)
         self.objectLabel.center = self.center
         self.objectLabel.textAlignment = .center
+        self.objectLabel.tag = 10
         self.addSubview(self.objectLabel)
         
+        // move UILabel to foreground
+        self.subviews.filter({$0 is UILabel}).forEach { label in
+            self.bringSubviewToFront(label)
+        }
     }
     
     @objc func tapGesture(gestureRecognizer: UITapGestureRecognizer) {
-        if (identifier != nil) {
-            let utterance = AVSpeechUtterance(string: identifier!) // 読み上げるtext
-            utterance.voice = AVSpeechSynthesisVoice(language: "en-US") // 言語
-            utterance.rate = 0.5; // 読み上げ速度
-            utterance.pitchMultiplier = 1.0; // 読み上げる声のピッチ(1.0でSiri)
-            utterance.preUtteranceDelay = 0.2; // 読み上げるまでのため
-            self.speechSynthesizer.speak(utterance)
-        }
-
         delegate?.tapGesture(identifier: self.identifier)
     }
 }
